requirements.txt
pbr
pygerrit2
GitPython
lxml
requests
